refactor: close database in cli packages
